+++ darwin/Classes/Music.swift
@@ -239,6 +239,19 @@ public class Player : NSObject, AVAudioPlayerDelegate {
         
         
         self.deinitMediaPlayerNotifEvent()
+        
+         // added seekBar 
+        if #available(iOS 9.1, *) {
+            commandCenter.changePlaybackPositionCommand.isEnabled = true
+            commandCenter.changePlaybackPositionCommand.addTarget { event in
+                if let event = event as? MPChangePlaybackPositionCommandEvent {
+                    let time = CMTime(seconds: event.positionTime, preferredTimescale: 1000000)
+                    self.player?.seek(to: time)
+                }
+                return .success
+            }
+        }
+        
         // Add handler for Play Command
         commandCenter.playCommand.isEnabled = (self.notificationSettings ?? NotificationSettings()).playPauseEnabled
         commandCenter.playCommand.addTarget { (event) -> MPRemoteCommandHandlerStatus in
@@ -347,7 +360,7 @@ public class Player : NSObject, AVAudioPlayerDelegate {
 
         
         if ((self.notificationSettings ?? NotificationSettings()).seekBarEnabled) {
-            self.nowPlayingInfo[MPMediaItemPropertyPlaybackDuration] = self.currentSongDuration
+            self.nowPlayingInfo[MPMediaItemPropertyPlaybackDuration] = self.currentSongDurationMs / 1000
             self.nowPlayingInfo[MPNowPlayingInfoPropertyElapsedPlaybackTime] = _currentTime / 1000
         } else {
             self.nowPlayingInfo[MPMediaItemPropertyPlaybackDuration] = 0
@@ -591,9 +604,10 @@ public class Player : NSObject, AVAudioPlayerDelegate {
                     }
                     
                     self?._playingPath = assetPath
-                    self?.setBuffering(false)
+                    //self?.setBuffering(false)
                     
                     self?.addPostPlayingBufferListeners(item: item)
+                    self?.addPlayerStatusListeners(item: (self?.player)!);
                     
                     result(nil)
                 case .failed:
@@ -658,9 +672,13 @@ public class Player : NSObject, AVAudioPlayerDelegate {
     }
     
     private func addPostPlayingBufferListeners(item : SlowMoPlayerItem){
-        observerStatus.append( item.observe(\.isPlaybackBufferEmpty, options: [.new]) { [weak self] (_, _) in
+        observerStatus.append( item.observe(\.isPlaybackBufferEmpty, options: [.new]) { [weak self] (value, _) in
             // show buffering
-            self?.setBuffering(true)
+            if(value.isPlaybackBufferEmpty){
+             self?.setBuffering(true)
+            }else{
+            self?.setBuffering(false)
+            }
         })
         
         observerStatus.append( item.observe(\.isPlaybackLikelyToKeepUp, options: [.new]) { [weak self] (_, _) in
@@ -672,6 +690,25 @@ public class Player : NSObject, AVAudioPlayerDelegate {
             // hide buffering
             self?.setBuffering(false)
         })
+    }
+    
+    
+    private func addPlayerStatusListeners(item : AVQueuePlayer){
+        if #available(iOS 10.0, OSX 10.12, *){
+            observerStatus.append( item.observe(\.timeControlStatus, options: [.new]) { [weak self] (value, _) in
+                // show buffering
+                if(value.timeControlStatus == AVPlayer.TimeControlStatus.playing){
+                    self?.playing = true;
+                    self?.updateNotifStatus(playing: true, stopped: false, rate: self?.player?.rate)
+                }else if(value.timeControlStatus == AVPlayer.TimeControlStatus.paused){
+                    self?.playing = false;
+                    self?.updateNotifStatus(playing: false, stopped: false, rate: 0)
+                }else{
+                    self?.playing = false;
+                    self?.updateNotifStatus(playing: false, stopped: false, rate: 0)
+                }
+            })
+        }
     }
     
     func getMillisecondsFromCMTime(_ time: CMTime) -> Double {
@@ -780,7 +817,6 @@ public class Player : NSObject, AVAudioPlayerDelegate {
         self.updateNotifStatus(playing: self.playing, stopped: true, rate: self.player?.rate)
         
         self.player?.seek(to: CMTime.zero)
-        self.player = nil   
         self.playing = false
         self.currentTimeTimer?.invalidate()
         #if os(iOS)
@@ -794,16 +830,21 @@ public class Player : NSObject, AVAudioPlayerDelegate {
         #if os(iOS)
         self.nowPlayingInfo.removeAll()
         #endif
+        self.player = nil
     }
     
     func play(){
-        self.player?.play()
-        self.player?.rate = self.rate
+        if #available(iOS 10.0, *) {
+            self.player?.playImmediately(atRate: self.rate)
+        } else {
+            self.player?.play()
+            self.player?.rate = self.rate
+        }
         self.currentTimeTimer = Timer.scheduledTimer(timeInterval: 0.3, target: self, selector: #selector(updateTimer), userInfo: nil, repeats: true)
         self.currentTimeTimer?.fire()
-        self.playing = true
+//        self.playing = true
         
-        self.updateNotifStatus(playing: self.playing, stopped: false, rate: self.player?.rate)
+        //self.updateNotifStatus(playing: self.playing, stopped: false, rate: self.player?.rate)
     }
     
     private var looper: Any?
@@ -861,7 +902,11 @@ public class Player : NSObject, AVAudioPlayerDelegate {
                 if(self.displayMediaPlayerNotification){
                     #if os(iOS)
                     self.nowPlayingInfo[MPNowPlayingInfoPropertyElapsedPlaybackTime] = _currentTime / 1000
+<<<<<<< HEAD
                     self.nowPlayingInfo[MPNowPlayingInfoPropertyPlaybackRate] = self.player!.rate / 1000
+=======
+                    self.nowPlayingInfo[MPNowPlayingInfoPropertyPlaybackRate] = self.player!.rate
+>>>>>>> upstream/master
                     MPNowPlayingInfoCenter.default().nowPlayingInfo = nowPlayingInfo
                     #endif
                 }
@@ -915,9 +960,9 @@ public class Player : NSObject, AVAudioPlayerDelegate {
     func pause(){
         self.player?.pause()
         
-        self.updateNotifStatus(playing: false, stopped: false, rate: 0)
+//        self.updateNotifStatus(playing: false, stopped: false, rate: 0)
         
-        self.playing = false
+//        self.playing = false
         self.currentTimeTimer?.invalidate()
     }
     
@@ -1124,7 +1169,7 @@ class Music : NSObject, FlutterPlugin {
                 guard let volume = args["volume"] as? Double else {
                     result(FlutterError(
                         code: "METHOD_CALL",
-                        message: call.method + " Arguments[volume] must be a String",
+                        message: call.method + " Arguments[volume] must be a Double",
                         details: nil)
                     )
                     break