+++ lib/src/PhoneStrategy.dart
@@ -17,20 +17,17 @@ class AudioFocusStrategy {
   final bool resumeAfterInterruption;
   final bool resumeOthersPlayersAfterDone;
 
-  /***
-   * Don't request focus
-   */
+  /// Don't request focus
   AudioFocusStrategy.none()
       : this.request = false,
         this.resumeAfterInterruption = false,
         this.resumeOthersPlayersAfterDone = false;
 
-  /***
-   * Request focus
-   *
-   * resumeAfterInterruption : When other app request focus (phone call, other player like spotify play), pause the AssetMediaPlayer
-   * resumeOthersPlayersAfterDone : When this player finish to play, tell others native players (like spotify), to resume
-   */
+  /// Request focus
+  ///
+  /// resumeAfterInterruption : When other app request focus (phone call, other player like spotify play), pause the AssetMediaPlayer
+  ///
+  /// resumeOthersPlayersAfterDone : When this player finish to play, tell others native players (like spotify), to resume
   const AudioFocusStrategy.request(
       {this.resumeAfterInterruption = false,
       this.resumeOthersPlayersAfterDone = false})